When dealing with many overloaded templates, include them in reverse order.
This is necessary due to the way Ruby's module inclusion works with
inheritance.  Including a module puts it at the base of the
inheritance chain.  For example, if a class hierarchy looks like
A->B->C, and C includes D, then it looks like A->B->D->C. This means
that the base template, which is last in the full_paths list, should
actually be included first, so that it ends up farthest back in the
inheritance chain.