Add support for @group and @endgroup
Comments must be in the form

    # @group NAME

To end a group use "# @endgroup". A group does not need to be ended for another to begin. If you want to start a new group, simply call a new @group.

Note that these tags are not treated as classical "tags" and cannot be overridden.