Add yardoc switches: `--tag`, `--type-tag`, `--type-name-tag`, `--name-tag`, `--title-tag` and `--hide-tag`
Allows the creation of custom YARD tags at the command line without requiring a YARD extension. Tags can be defined in the following way:

    $ yardoc --tag tagname:"Title of Tag"

Similarly a "typed", "named" or "titled" tag (or a few combinations) can be added with the respective switches.

Note the ':' separating the tag's @tagname from the title used in templates. The quotations is used to escape spaces in the shell (quotations should not be used when done inside Ruby). Titles are optional (omit the ":TITLE" part).

Tags added in this way are automatically displayed in templates in the order that they are defined with the title given. If a given tag should not be displayed in the templates, use the `--hide-tag` switch to hide it:

    $ yardoc --tag tagname --hide-tag tagname

This changeset also makes Library.visible_tags into an attribute.