Refactor 'sections' template command to take Section objects and build tree-like structures instead of plain Arrays.
This change replaces all sections with Section objects, allowing subsections to be accessed with Section#[] by name:

    def init
      sections :a, :b, :c, [:sub1, :sub2]
      sections[:c].place(:sub3).after(:sub1)
      puts "(sections is now [:a, :b, :c, [:sub1, :sub2, :sub3]])"
    end

This removes the need to use indexes, which is problematic when multiple plugins overload the same template (because the numeric index is not guaranteed).

This change also adds Insertion#before_any and Insertion#after_any to insert an object before any section no matter how deeply nested it is.

    [1,[2,[3]]].place(4).after_any(3) #=> [1, [2, [3, 4]]]

The above template could be refactored as:

    def init
      sections :a, :b, :c, [:sub1, :sub2]
      sections.place(:sub3).after_any(:sub1)
    end

Closes gh-58