Add transitive tags (and --transitive-tag). Transitive tags allow a user to declare a tag on a namespace (like a module or class) and have the tag be automatically applied to all methods and objects inside the namespace.
Consider declaring "@since" on a new class. Instead of defining @since on every method inside the new class, a user can simply declare it on the class itself and the tag will automatically be applied to the methods inside. For example:

    # @since 1.1
    class Foo
      def bar; end
    end

The Foo#bar method will have a @since tag defined as well.