Add "(see PATH)" syntax for Docstrings. Allows the ability to reference an entire docstrings (including tags) using the following syntax:
    # (see OtherObject)
    class Object; end

    # Real doctring here
    # @since 0.1
    class OtherObject

The above example will use the docstring on OtherObject on both objects (including tags).

The (see ...) syntax must be at the start of the docstring, though extra tags/info is allowed to be added afterwards. The following is valid:

    # (see OtherObject)
    # extra info
    # @author Name

This change also adds Docstring#+ to add two docstrings together and maintain tag information.