Refactored configuration support, adding YARD::Config class. Plugins are no longer loaded automatically by default, but this can be changed by editing ~/.yard/config as detailed below:
The Config class now manages the loading of the ~/.yard/config file, a YAML
file containing the YARD::Config::DEFAULT_CONFIG_OPTIONS options. It also
manages the loading of plugins.

The configuration file has 3 options specifying rules to use when loading plugins:

    :load_plugins (default: false) When enabled, gem plugins are automatically
        loaded (0.5.x and 0.6.0/1 behaviour)
    :autoload_plugins specifies a list of plugins to load automatically
    :ignored_plugins specifies a list of plugins to ignore

The ~/.yard/ignored_plugins file is now deprecated in favour of specifying
the :ignored_plugins setting in the config file.