Use remove_method instead of undef_method
undef_method has very different semantics. It prevents the class
from responding to the named method. This means that also original
methods on super classes won't be called. This for example results
in weird behavior on Rubinius since the original method_missing
method can't be called which normally throws the NoMethodError.

Also see:
http://www.nach-vorne.de/2008/2/28/undef_method-remove_method/index.html