Type list ends at next occurrence of matching delimiter.
With this patch, the type list ends at the next occurrence of the closing
delimiter corresponding to the opening delimiter.  For example, if the type
list begins with `[`, then it ends at the next `]`.  If you want to include `]`
in the type list, then wrap the type list in a different delimiter pair:

    # @param {[Integer, Integer]} point
    #           co-ordinates in array