Only add /version/ to URLs accessed with /version/
Adds request.version_supplied to detect whether the version was parsed out
of the URL. Rewrote some of the URL helpers and serializers to respect this
new behaviour.

The biggest change is that the DocServerSerializer will no longer add the
/prefix to the serialized_path, and it will now ignore the command passed
in to the constructor (which is now optional). Instead,
DocServerHelper#url_for should be used to generate a URL with a prefix.

Closes #426