do not assume that html_syntax_highlight always gets called from htmlify
By having moved up the language detection and removing of !!!LANG from
the input source we broke html_syntax_highlight when it was being
called independently, as it wouldn't sanitize input nor have a
fallback language.

We now have to detect the language twice, once for writing the correct
<pre> tag and once to ensure that html_syntax_highlight is working
with sane inputs.