pot: add "yard i18n" command that generates .pot
YARD::CLI::I18n inherits YARD::CLI::Yardoc because "i18n" command
use the same options in "doc" command. But it will be better that
we share the same options in another way because "i18n" command
doesn't need all options in "doc" command.

We need to discuss how to implement it. So I use inheritance for now.