Add experimental {include:Object attribute} syntax to expand attribute values from code objects.
Attribute can also be @tagname or @tagname[N], which would expand the Nth
tag by that name on the object. In such a case, an extra attribute can
be provided to access the tag attribute, for example, to expand the
3rd param name of an object:

  {include:#method param[2] name}

To get the value of a constant:

  {include:CONST value}

By default, the attribute used is "docstring", which expands the full
docstring of the object.