Support for custom reference links for CodeObjects
The feature allows the registration of arbitrary "link keywords" that
can be used to link against certain object types. This is handy when
defining custom CodeObject implementations where referencing them by
path is unnatural and not human readable.

Modifications:
  * RegistryStore now tracks object_links which are a keyword ->
    object_type map
  * Proxy objects can now act as "links" and resolve themselves
  * HTMLHelper#link_object handles linkable Proxy objects

There are specs that cover the added methods, and all
the modifications are documented with proper YARD syntax.