JavaScript bug fix for the default template.
When using the frames view in the default template,
it attempts to set the page title when you load a page in the
main frame.  If you are serving content from disk using file://,
some browsers, like Chrome, will raise an error. This stops
the rest of the JavaScript on the page from running.

This pull request adds a try-catch block around the window title
assignment statement.  In the case when this fails, the title simply
is not updated.