Add CONSTANTSTART regular expression
Prior to this commit, the regular expression used in the proxy_path
method was hard coded. In addition to be inconsistent with the rest
of the regular expressions in CodeObjects::Proxy (which use the constants
defined in CodeObjects::Base), this also meant that it was not possible
to patch the regular expressions to accommodate namespaces in other
languages.

In order to fix these issues, add a CONSTANTSTART regex constant that
matches the first letter of constants that begin with an uppercase letter.