[config_dir] use one constant and fallback to /tmp
This lets applications launch when ENV["HOME"] is unset, blank, or
relative.

For a variety of reasons, when launching my app, HOME is empty
and gems I need depend on YARD. I believe we shouldn't prevent
apps from starting in this case.

This cleans up our usage of "~/.yard" in a few places to reference the
same constant, which falls back to "/tmp.yard" when HOME is unset.

This also removes the deprecated ::YARD::CONFIG_DIR.
It was deprecated Sep 16, 2010
598b051202549122dc12411ce6e7b26b4c98d66b