Fix unexpected writers in --list-undoc.
A writer documented by the @!attribute directive with the "rw" option
is listed in --list-undoc even if the attribute is documented.
This is because the MethodObject object of the writer always
has an empty docstring.

With this fix, the MethodObject objects of the reader and the
writer created by the @!attribute attribute share a docstring.