Matching of override comments to methods defined in C extension is stricter
Previously, an override comment for `Foo#baz` would also be applied to `Bar#baz`.
This was especially problematic, because if an extension defined methods with the
same names on two different classes, the override comment which appeared *first*
in the source code would always be favored over a more specific comment which
appeared later.