Bump RSpec to 3.3
Now we need to fix all the breakage!

At the same time, examples which are named like: 'it "should ..."' will be renamed
to eliminate the redundant "should". It is just an extra word which is repeated
in the name of every single example and doesn't communicate anything. Brevity is
enhanced by just stating directly what the correct code does, not what it "should"
do.