FileSystemSerializer now maps objects to case insensitive filenames.
The serializer object must be initialized AFTER the registry is
finalized in order for mapping to work correctly. Note also
that generated filenames of colliding objects may not be
consistent across runs of YARD and are order dependent
(Foo may be Foo.html or Foo2.html) and should not be relied on
for URL structure.

Fixes #906