Make FileSystemSerializer#serialized_path spec robust on MRI 1.8.7
More nonsense with different interpreter versions! Since Hashes in Ruby 1.8.7
do not have a defined iteration order, when FileSystemSerializer serializes multiple
code objects whose names would clash on a non-case-sensitive filesystem, it may
append the disambiguating "2" onto either object's name.

On 1.9+, it will always append the "2" to the name of the code object which was
initialized later.