Allow wrapping extra file meta data in HTML comment
Meta data for extra files has to be placed at the top of the
file. Since each line has to be prefixed with a `#` those lines are
displayed as headers when viewing markdown files on GitHub. This
commit allows hiding the meta data section in rendered markdown on
GitHub by wrapping it in an HTML comment:

    <!--
    # @title Some title
    -->

    Rest of document