Resolve macros through alias names
Allows aliases of attached macros to be used in
DSL handlers. This now works:

    class Foo
      # @!macro foo
      #   Documentation
      def self.foo(x) end

      # Alternative name for foo
      class << self; alias bar foo; end

      bar :test # Note alias usage
    end

This change does not support defining macros on
alias declarations if they already have a macro
on the original object. This behavior is currently
unsupported, but might be in the future.

This change also adds a subtle behavior difference to aliases.
Aliases no longer attempt to parse directives from the original
method. This caused a bug whereby macros and directives would
be re-executed on the alias and redefine macros. It is possible
to disable macro parsing explicitly on alias declarations, but
this is not yet implemented. Once this is implemented, old
alias behavior could be re-introduced.

Fixes #941