DSL handler won't create method if attached macro has directives
YARD will no longer attempt to create a method object for a DSL
style declaration with code comments if the declaration has an
attached macro with directive text. If directives are used in
an attached macro, you must opt-in to @!method or @!attribute
to create the code object yourself. If you do not use a directive,
the DSL handler will create a method for you. The following
two scenarios now work:

    class Foo
      # @!macro attach
      #   Docstring for $1
      def self.attach(x) end

      attach :bar # creates method Foo#bar

      # @!macro attach2
      #   @!method macro_$1(x)
      def self.attach2(x) end

      attach2 :baz # creates method Foo#macro_baz
    end

The following will NO LONGER create a method. This was previously
undefined behavior (since macros themselves are not meant to create
objects):

    class Foo
      # @!macro attach
      #   A documentation string
      #   @!visibility private
      def self.attach(x) end

      attach :bar

      # Foo#bar has not been created.
      # The attach macro has a visibility directive,
      # but no @!method directive to create the object.
      # YARD will set visibility, but nothing else.
    end

Fixes #661