Disallow type specification to begin after newline
This change disallows the following tag formatting to minimize
ambiguity when differentiating tag types from body text:

    @param name
      [type, specification, here] docstring here
    @param
      [other, type, specification] name docstring here

Supporting the above two examples make it impossible to distinguish
reftags from directives:

    @!attribute
      (see Foo)

Because (see Foo) would otherwise be parsed as a type specification.

If an opening character is detected before a newline, YARD will parse
the specification completely. In other words, the following behavior
is still supported:

    @param name [Long, Type, Specification,
      Here] and more docstring text here

Or even:

    @param long_parameter_name [
      Type, Specification, Here] docstring

This is a breaking change but is unlikely to have wide impact
due to the nature of the formatting. It is possible to implement
this in a non-breaking way by only enabling this parsing style
on directives, but the extra complexity is not seen as worth it
given the extremely low impact.

Fixes #726