Add modern rspec skeleton configuration to spec_helper
The latest releases of rspec include a number of cool options which are
not on by default to preserve backwards compatibility. See
http://ivoanjo.me/blog/2016/09/25/why-you-should-regenerate-your-spec-helper/
for more details and a short explaination for each.

For now, I've kept disabled the following options due to test failures:
* `mocks.verify_partial_doubles`
* `config.disable_monkey_patching!`
* `config.warnings`
* `config.order = :random`

I'll tackle these over the next commits.