Fix test example
This test example has been introduced in 523256605c04376a9961a347,
and is expected to cover situation when processed document includes
some byte sequence which is invalid in given encoding.  However,
due to a subtle bug, the tested string was not satisfying this
requirement, and thus example setup was incorrect.

This commit changes string encoding to UTF-8, in which it has invalid
sequence of bytes, which is expected.  Also, yet another assertion is
added in order to prevent this bug from being re-introduced.