Ignore all magic comments at the start of file
YARD previously recognized three magic comments:
- Shebang
- Encoding
- Frozen String Literal

Standard Ruby also comes with another magic comment: `warn_ident`.

The use of magic comments is not limited to what comes built-in with
Ruby, however; some tooling uses magic comments as configuration.

It is desirable for YARD to ignore all magic comments at the beginning
of the file, in order to not clobber module documentation with magic
comments.