Do not highlight Ruby snippets twice
The #html_markup_ruby method, which is responsible for creating HTML from
Ruby snippets, produces output with syntax highlighting already applied.
Doing it for the second time messes things up.

It wasn't a big deal before the regular expression in #parse_codeblocks
has been relaxed.  Only a slightly different set of HTML attributes was
preventing the rendered Ruby listings from being matched.  But now,
a proper fix is needed, and it is introduced in this very commit.